### 2024-10-15 #7152 @H-Snoussi
Hello Beatriz,

Unfortunately, this is outside my expertise.

Best regards,
Haykel

On Tue, Oct 15, 2024 at 3:49 AM Beatriz Costa Gomes <
***@***.***> wrote:

> Hi 👋 @ianhi <https://github.com/ianhi> @jrkerns
> <https://github.com/jrkerns> @H-Snoussi <https://github.com/H-Snoussi> -
> would you be available and interested in reviewing this submission for the
> Journal of Open Source Software?
>
> SLAP2 Utils: Tools for Processing SLAP2 Data
> <https://github.com/openjournals/joss-reviews/issues/7152>
> #7152 <https://github.com/openjournals/joss-reviews/issues/7152>
>
> Thanks in advance!
>
