### 2024-10-20 #7152 @Kevin-Mattheus-Moerman
@openjournals/dev an editor has been assigned here but this one is still showing up on the dashboard as unassigned. Can you check what is happening? Thanks

@mooniean this does not influence anything here.
