### 2024-10-22 #7152 @xuanxu
> an editor has been assigned here but this one is still showing up on the dashboard as unassigned.

fixed!
