### 2024-10-28 #7152 @mooniean
Hi @ThFriedrich ! Would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> SLAP2 Utils: Tools for Processing SLAP2 Data

https://github.com/openjournals/joss-reviews/issues/7152

Thanks in advance!
