### 2024-10-28 #7152 @H-Snoussi
Again, please remove my email from this loop. Thanks.

On Mon, Oct 28, 2024 at 11:52 AM Beatriz Costa Gomes <
***@***.***> wrote:

> Hi @ThFriedrich <https://github.com/ThFriedrich> ! Would you be available
> and interested in reviewing this submission for the Journal of Open Source
> Software?
>
> SLAP2 Utils: Tools for Processing SLAP2 Data
>
> #7152 <https://github.com/openjournals/joss-reviews/issues/7152>
>
> Thanks in advance!
>
