### 2024-10-28 #7152 @danielskatz
@H-Snoussi - you can go to the GitHub page for this issue (see the bottom of the email) and then turn off notifications (see the image below which is the right side of the page, which show the unsubscribe from notifications button). No one else can do this for you - sorry.
