### 2024-11-15 #7152 @Peter-Hogg
@ThFriedrich, we have a sample data file in the repo's example folder. We are always happy to provide more data as needed.
