### 2024-11-19 1737c7a
Update gratingStimScript.m
