### 2024-11-19 2cc52bb
Update GenGrating.m (added duration of presentation as a parameter)
