### 2024-11-19 665fe5c
Update GenGrating.m
