### 2024-11-20 15722cb
Update gratingProbing.m (finalized)
