### 2024-12-04 3d95c50
implement getLineData
