+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "SLAP2_UTILS"
-version = "0.3.0"
+version = "0.3.5"
 authors = [
   { name="Peter Hogg", email="peter.hogg@ubc.ca" },
   { name="Jerry Tong", email="jerrytong0810@gmail.com" }
