### 2025-01-03 #7152 @mooniean
Happy new year everyone!

@ianhi how is your availability currently - would you be up for this?
