### 2025-03-07 b925800
add compatibility for updated SLAP2 metadata format
