### 2025-04-14 #7152 @Kevin-Mattheus-Moerman
@peterhys, @Armavica,, @ianhi, @untzag would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled _SLAP2 Utils: Tools for Processing SLAP2 Data_ and features Python for processing of a custom microscopy file format.

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/Peter-Hogg/SLAP2_Utils) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07152/joss.07152/10.21105.joss.07152.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of an ongoing review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!
