### 2025-04-15 #8054 @editorialbot
[REVIEW]: SLAP2 Utils: Tools for Processing SLAP2 Data
**Submitting author:** @peter-hogg (<a href="http://orcid.org/0000-0003-2176-4977">Peter William Hogg</a>)
**Repository:** https://github.com/Peter-Hogg/SLAP2_Utils
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.0
