### 2025-04-17 5bdcd1e
getLineData: fix bug when saving only 1 channel
