+++ setup.cfg
@@ -1,4 +1,4 @@
 # Inside of setup.cfg
 [metadata]
 description-file = README.md
-version = 0.3.6
+version = 0.3.9
