+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "SLAP2_UTILS"
-version = "0.3.9"
+version = "0.4.0"
 authors = [
   { name="Peter Hogg", email="peterhogg2006@gmail.com" },
   { name="Jerry Tong", email="jerrytong0810@gmail.com" }
