### 2025-04-18 c059452
fix some bugs when reading in data
