### 2025-04-28 #8054 @jrkerns
I'm out of office currently but will get started when I get back!

On Mon, Apr 28, 2025, 4:42 AM Kevin Mattheus Moerman <
***@***.***> wrote:

> *Kevin-Mattheus-Moerman* left a comment (openjournals/joss-reviews#8054)
> <https://github.com/openjournals/joss-reviews/issues/8054#issuecomment-2834467326>
>
> @larsenkg <https://github.com/larsenkg>, @jrkerns
> <https://github.com/jrkerns> thanks for agreeing to review here.
>
> @larsenkg <https://github.com/larsenkg> I see you have already started.
> Feel free to comment here if you feel the authors should work on anything
> in particular.
>
> @jrkerns <https://github.com/jrkerns> are you able to get started too?
>
