### 2025-05-06 Peter-Hogg/SLAP2_Utils#2 @larsenkg
Cannot install package in Windows 11 with Python 3.12
I am running into a problem with installing the package. I am using Python 3.12 on Windows 11.

(I did not run into any problems installing this package in WSL2 with Python 3.10.)

# Steps to Reproduce
1. Create new virtual environment: `> python -m venv venv`
2. Activate virtual environment: `> .\venv\Scripts\activate`
3. Install from source or from pypi:
   - `> pip install SLAP2-UTILS`, or:
   - `> git clone https://github.com/Peter-Hogg/SLAP2_Utils.git` then `> pip install .`

These steps lead to Error 1 below. I thought to try updating/reinstalling `setuptools` with `> pip install setuptools`, which leads to Error 2. The second error complains about `ImpImporter`, which was deprecated in Python 3.12. If this package doesn't support Python 3.12, a note in the README would be valuable. Otherwise, I anticipate many users will run into this problem.

# Error 1
```
(venv) PS C:\Users\larsenkg\projects\joss\slap2_pip> pip install SLAP2-UTILS
Collecting SLAP2-UTILS
  Using cached slap2_utils-0.4.0-py3-none-any.whl.metadata (3.7 kB)
Collecting numpy==1.23.5 (from SLAP2-UTILS)
  Using cached numpy-1.23.5.tar.gz (10.7 MB)
  Installing build dependencies ... done
ERROR: Exception:
Traceback (most recent call last):
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\cli\base_command.py", line 105, in _run_wrapper
    status = _inner_run()
             ^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\cli\base_command.py", line 96, in _inner_run
    return self.run(options, args)
           ^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\cli\req_command.py", line 68, in wrapper
    return func(self, options, args)
           ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\commands\install.py", line 387, in run
    requirement_set = resolver.resolve(
                      ^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\resolver.py", line 96, in resolve
    result = self._result = resolver.resolve(
                            ^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\resolvelib\resolvers\resolution.py", line 515, in resolve
    state = resolution.resolve(requirements, max_rounds=max_rounds)
            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\resolvelib\resolvers\resolution.py", line 444, in resolve
    failure_criterion = self._attempt_to_pin_criterion(name)
                        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\resolvelib\resolvers\resolution.py", line 211, in _attempt_to_pin_criterion
    criteria = self._get_updated_criteria(candidate)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\resolvelib\resolvers\resolution.py", line 202, in _get_updated_criteria
    self._add_to_criteria(criteria, requirement, parent=candidate)
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\resolvelib\resolvers\resolution.py", line 141, in _add_to_criteria
    if not criterion.candidates:
           ^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\resolvelib\structs.py", line 194, in __bool__
    return bool(self._sequence)
           ^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\found_candidates.py", line 163, in __bool__
    self._bool = any(self)
                 ^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\found_candidates.py", line 147, in <genexpr>
    return (c for c in iterator if id(c) not in self._incompatible_ids)
                       ^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\found_candidates.py", line 37, in _iter_built
    candidate = func()
                ^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\factory.py", line 187, in _make_candidate_from_link
    base: Optional[BaseCandidate] = self._make_base_candidate_from_link(
                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\factory.py", line 233, in _make_base_candidate_from_link
    self._link_candidate_cache[link] = LinkCandidate(
                                       ^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\candidates.py", line 306, in __init__
    super().__init__(
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\candidates.py", line 159, in __init__
    self.dist = self._prepare()
                ^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\candidates.py", line 236, in _prepare
    dist = self._prepare_distribution()
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\resolution\resolvelib\candidates.py", line 317, in _prepare_distribution
    return preparer.prepare_linked_requirement(self._ireq, parallel_builds=True)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\operations\prepare.py", line 532, in prepare_linked_requirement
    return self._prepare_linked_requirement(req, parallel_builds)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\operations\prepare.py", line 647, in _prepare_linked_requirement
    dist = _get_prepared_distribution(
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\operations\prepare.py", line 71, in _get_prepared_distribution
    abstract_dist.prepare_distribution_metadata(
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\distributions\sdist.py", line 56, in prepare_distribution_metadata
    self._install_build_reqs(finder)
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\distributions\sdist.py", line 126, in _install_build_reqs
    build_reqs = self._get_build_requires_wheel()
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\distributions\sdist.py", line 103, in _get_build_requires_wheel
    return backend.get_requires_for_build_wheel()
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_internal\utils\misc.py", line 702, in get_requires_for_build_wheel
    return super().get_requires_for_build_wheel(config_settings=cs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\pyproject_hooks\_impl.py", line 196, in get_requires_for_build_wheel
    return self._call_hook(
           ^^^^^^^^^^^^^^^^
  File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\pyproject_hooks\_impl.py", line 402, in _call_hook
    raise BackendUnavailable(
pip._vendor.pyproject_hooks._impl.BackendUnavailable: Cannot import 'setuptools.build_meta'
```

# Error 2
```
(venv) PS C:\Users\larsenkg\projects\joss\slap2_pip> pip install SLAP2-UTILS
Collecting SLAP2-UTILS
  Using cached slap2_utils-0.4.0-py3-none-any.whl.metadata (3.7 kB)
Collecting numpy==1.23.5 (from SLAP2-UTILS)
  Using cached numpy-1.23.5.tar.gz (10.7 MB)
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [33 lines of output]
      Traceback (most recent call last):
        File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\pyproject_hooks\_in_process\_in_process.py", line 389, in <module>
          main()
        File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\pyproject_hooks\_in_process\_in_process.py", line 373, in main
          json_out["return_val"] = hook(**hook_input["kwargs"])
                                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\pyproject_hooks\_in_process\_in_process.py", line 137, in get_requires_for_build_wheel
          backend = _build_backend()
                    ^^^^^^^^^^^^^^^^
        File "C:\Users\larsenkg\projects\joss\slap2_pip\venv\Lib\site-packages\pip\_vendor\pyproject_hooks\_in_process\_in_process.py", line 70, in _build_backend
          obj = import_module(mod_path)
                ^^^^^^^^^^^^^^^^^^^^^^^
        File "C:\Program Files\WindowsApps\PythonSoftwareFoundation.Python.3.12_3.12.2800.0_x64__qbz5n2kfra8p0\Lib\importlib\__init__.py", line 90, in import_module
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "<frozen importlib._bootstrap>", line 1387, in _gcd_import
        File "<frozen importlib._bootstrap>", line 1360, in _find_and_load
        File "<frozen importlib._bootstrap>", line 1310, in _find_and_load_unlocked
        File "<frozen importlib._bootstrap>", line 488, in _call_with_frames_removed
        File "<frozen importlib._bootstrap>", line 1387, in _gcd_import
        File "<frozen importlib._bootstrap>", line 1360, in _find_and_load
        File "<frozen importlib._bootstrap>", line 1331, in _find_and_load_unlocked
        File "<frozen importlib._bootstrap>", line 935, in _load_unlocked
        File "<frozen importlib._bootstrap_external>", line 999, in exec_module
        File "<frozen importlib._bootstrap>", line 488, in _call_with_frames_removed
        File "C:\Users\larsenkg\AppData\Local\Temp\pip-build-env-vayy_q2e\overlay\Lib\site-packages\setuptools\__init__.py", line 16, in <module>
          import setuptools.version
        File "C:\Users\larsenkg\AppData\Local\Temp\pip-build-env-vayy_q2e\overlay\Lib\site-packages\setuptools\version.py", line 1, in <module>
          import pkg_resources
        File "C:\Users\larsenkg\AppData\Local\Temp\pip-build-env-vayy_q2e\overlay\Lib\site-packages\pkg_resources\__init__.py", line 2172, in <module>
          register_finder(pkgutil.ImpImporter, find_on_path)
                          ^^^^^^^^^^^^^^^^^^^
      AttributeError: module 'pkgutil' has no attribute 'ImpImporter'. Did you mean: 'zipimporter'?
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

```
