### 2025-05-06 #8054 @larsenkg
I was able to successfully install this package in WSL2 with Python 3.10, but have not yet been able to install it on Windows 11 with Python 3.12. I filed an issue here: Peter-Hogg/SLAP2_Utils#2

I will continue the review using the WSL2 installation.
