### 2025-05-06 Peter-Hogg/SLAP2_Utils#2 @Peter-Hogg
The requirement for numpy==1.23.5 is causing the issue. However, I've been testing newer versions of Numpy and everything works fine in the example notebook. I had that numpy requirement for personal preference. I'll remove it; then it should install for Python 3.12 without problems.
