### 2025-05-06 Peter-Hogg/SLAP2_Utils#2 @Peter-Hogg
The new release fixes the issue for Python 3.12 when I tested locally
