### 2025-05-14 Peter-Hogg/SLAP2_Utils#2 @larsenkg
I can confirm that the latest release works for my on Python 3.12. Thank you!
