### 2025-05-14 #8054 @larsenkg
I have completed my review. I believe this paper should be published as it represents the only software capable of interacting with the complex data structure required for SLAP2 measurements.

I have a few minor concerns:

1. The authors claim that this software package "provide[s] extensions to the SLAP2 microscope control and image acquisition software". I cannot verify the functionality of the software features relating to this claim. No specific action is required, though please see item 3 below.
2. The repository contains an "experiments" directory which includes specific experiments that are not necessarily intended for the users of `SLAP2_Utils`. These are the only files that I've noticed that are not documented. I recommend cleaning up this directory, adding documentation where necessary, and (perhaps) adding example experiments to the "examples" directory.
3. The example notebook shows basic use cases using (I believe) simulated data. It may be beneficial to users to show a real analysis problem, though I imagine that may not be possible. At the very least, I recommend creating another example showing how to use the microscope control features (see item 1 above).
4. The only automated test appears to test the functionality of reading SLAP2 data. This seems sparse to me, considering the other features this software provides. I won't mandate a specific number of tests, but I do recommend trying to expand test coverage because it will only help you (and any future maintainers) down the road.

I have left three check boxes unchecked while we discuss these points. If the consensus is that the existing examples and tests are sufficient, then I am willing to check them off as well.

Please let me know if you have any questions about these comments.
