### 2025-05-19 Peter-Hogg/SLAP2_Utils#3 @jrkerns
JOSS Review
This is a review for the submitted JOSS paper with items found during the review. The JOSS checklist item is listed along with the note.

- "Functionality: Have the functional claims of the software been confirmed?" In my opinion only some of the claims could be verified. Reading the SLAP file and associated metadata was accomplished through the provided example notebook. There is also a mention of a "Utility" UI. There was no documentation I found for this usage, so I could not verify this claim.
- "Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution." There is a clear list of dependencies. However, some seem superfluous and it's not clear the canonical list of packages. In `requirements.txt`, `notebook` is listed as a dependency, but I do not see it's usage anywhere in the library. `h5py`, `PyQT6`, `matplotlib`, `vispy`, `scikit-image`, `scipy` are all listed twice. `pyproject.toml` lists `pyneurotrace` but is not listed in `requirements.txt`. There is both a `pyproject.toml` and `setup.py` file. It's not clear which one is used for the PYPI package (which determines the dependencies at install time). There is a small formatting error for the cupy installation. A new blank like needs to separate the code block https://github.com/Peter-Hogg/SLAP2_Utils/blob/a29f941d76bc8bf9ef28248510bfb17d103b510c/docs/source/installation.rst?plain=1#L29-L30. There also appears to be expected imports that are not stated. E.g. [`zarr`, `dask`](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/imagestacks.py#L2-L9), [`apoc`](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/autoROI.py#L2), [`napari`](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/autoROI.py#L8), and many others. Furthermore, without pinned requirements in `requirements.txt`, there are possible pitfalls as some functions appear deprecated. E.g. [`skeletonize_3d`](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/autoROI.py#L12) is no longer a function in [`skimage.morphology`](https://scikit-image.org/docs/stable/api/skimage.morphology.html#skimage.morphology.skeletonize).
- "Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)." There is an example notebook buried in the `examples` directory. However, I saw no mention of this in the claimed canonical documentation. Note that you can [include notebooks](https://docs.readthedocs.com/platform/stable/guides/jupyter.html) in Sphinx documentation. It was unclear if the example .dat file was a "real-world" image. I found no documentation related to the UI or its usage in the documentation. If that section is still in progress I would remove it from the paper. It's unclear if [this](https://github.com/Peter-Hogg/SLAP2_Utils/tree/main/slap2_utils/dataviewer/plotWindows) is the intended module for usage.
- "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?" Documentation exists. However, the docs have errors in their builds. E.g. the `imagestacks` module does not populate because of a [build error](https://github.com/Peter-Hogg/SLAP2_Utils/actions/runs/14868419567/job/41750870790). I would encourage the authors to use `sphinx-build` as it provides a number of [helpful parameters](https://www.sphinx-doc.org/en/master/man/sphinx-build.html) such as `-n`, and `-W`. While quite nitty, many words that appear meant to be in monospace code are italicized due to single ticks vs double. E.g. [here](https://github.com/Peter-Hogg/SLAP2_Utils/blob/a29f941d76bc8bf9ef28248510bfb17d103b510c/docs/source/modules/acquisitioncontainer.rst?plain=1#L4).
- "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?" 1 test exists for reading a single datafile and metadata. It ran and passed. I did not find any testing for any other module or function. Coverage of the `slap2_utils` directory with the test was 13%. In my opinion, this is not acceptable for a package meant to be distributed. E.g. a simple name error in the [`adaptive_thresh`](https://github.com/Peter-Hogg/SLAP2_Utils/blob/a29f941d76bc8bf9ef28248510bfb17d103b510c/slap2_utils/filters/filters.py#L77) function uses `np` but numpy is not imported, meaning doing `filters.adaptive_thresh(...)` will always result in a `'np' is not defined` error.
- "A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?" A statement of need exists. The target audience is not directly addressed.
- "Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?" There are typos and grammatical issues throughout. E.g. line 65 "numpyr" -> "numpy", line 65/66 "memap" -> "[memmap](https://numpy.org/doc/stable/reference/generated/numpy.memmap.html)". OOP and FOV are abbreviated in parentheses but never used again. The definition of "SLAP" appears different between the paper and the repo. In the paper it's "Scanned Line Angular Projection" but in the [readme](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/README.md?plain=1#L4) it's "Scanned Line Projection Microscopy". "Utility" is misspelled in the section title, line 83.

General notes
- The name of the tool in the title of the paper is "SLAP2 Utils" and is referred to as such in the paper, however, the package itself is `SLAP2-UTILS`, and the repo name is "SLAP2_Utils" and that is used throughout the readme file. This is somewhat inconsistent. It would be nice if the package name in the paper matched the Python package which matched the name stated in the readme.
- The line between personal experiment repo and distributable library appears non-existent. There is an `experiments` module with what appears to be auto-generated, personalized files? https://github.com/Peter-Hogg/SLAP2_Utils/blob/a29f941d76bc8bf9ef28248510bfb17d103b510c/slap2_utils/experiments/haasSlap2UI.m#L8
