### 2025-05-19 #8054 @jrkerns
I have noted several items from the checklist in a [separate ticket](https://github.com/Peter-Hogg/SLAP2_Utils/issues/3) in the target repo. Overall, I find the submission to be below-average quality. @larsenkg has already noted some of these items. I'm sure the library accomplished its purpose for the primary authors. If they developed this as an ancillary project to their primary purpose of analyzing the images for their project, great. But it does not appear to be in a robust state _meant for peer usage_ at this time. If this was a normal scientific journal I would reject it but I'm not sure if JOSS' policy allows for this. I will leave that to the Editor. E.g. in the JOSS review article the [guiding principles](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#guiding-principles) state: "We like to think of JOSS as a ‘developer friendly’ journal. That is, if the submitting authors have followed best practices (have documentation, tests, continuous integration, and a license) then their review should be rapid." While many issues in this project are typical and fixable, the lack of meaningful tests, the plethora of rudimentary and import errors, and sparse documentation is one I can't work around from a Python user and library maintainer perspective.
