+++ paper/paper.md
@@ -39,13 +39,13 @@ To overcome the inherent speed limitations of laser scanning two-photon imaging,
 
 # SLAP2 Datafile Pipeline
 
-`SLAP2 Utils` provides an open-source version of Matlab functions [@mbfSLAP2] designed to interface with binary files generated from high-speed recordings of neural activity from the SLAP2 two-photon microscope. This implementation uses object-orientated programming (OOP), creating Python Classes for parsing out the data for analytical pipelines (see Figure 1 ) [@mbfSLAP2]. This is needed because, unlike raster imaging that the SLAP2 microscope can also preform, which generates TIF formatted digital images, high-speed random-access imaging produces two files: a binary file (.dat) with an accompanying metadata (.meta) file that stores acquisition information using the HDF5 format. The binary file consists of a header, which stores core acquisition settings in unsigned 32-bit integers, followed by the digitized recording of the microscope’s detectors in unsigned 16-bit integers. The `MetaData` file contains more detailed information about the acquisition, including key parameters, such as the `ParsePlan`, a collection of data describing the timing of axial focus and coordinates being sampled   by the microscope. This information about the acquisition is needed to index the recording in the binary file. During random-access imaging, the microscope will sample ROIs at different acquisition rates in an order determined by a planning algorithm, which may not be in any particular spatial order. Given the complications in random-access imaging and the organization of the recording in the binary file, a simple linear readout of the DAT file’s contents would be insufficient to extract the neural activity. The core function of our Python library handles this task with our `DataFile` object that parses raw data and meta-data stored in these files directly into a Python environment. This `DataFile` class, upon initialization, loads information from the file header of the binary file and creates a subclass, MetaData, which will load additional acquisition information using the h5py library [@h5py]. The digitized signals of the acquisition are not loaded into memory. Rather, the Datafile class generates a memory map (memap) of the 16-bit raw recording in the binary file using Numpyr[@numpy]. This memap is used to index specific data in the raw recording when needed, helping end users work with larger datasets collected at kilohertz acquisition rates. 
[…]
