### 2025-05-20 4f78c33
Stop nb from running
