+++ _sources/SLAP2_Utils_example.ipynb.txt
@@ -1,0 +1,582 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "0f87eb4d-b0fd-401d-ac6e-92042583292f",
+   "metadata": {},
+   "source": [
+    "## SLAP2 Utils Example Notebook\n",
+    "\n",
+    "This notebook will highlight core features included in `SLAP2_Utils`. The example will demonstrate how to extract useful metadata, extract a trace signal, and use basic plotting. This will give users a better understanding of how to work with the library with their own data. For further information on any functions used in the notebook, please reference the [documentation](https://peter-hogg.github.io/SLAP2_Utils/)\n"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "12af0fa7-b0e2-4f52-8372-367931141ce9",
+   "metadata": {},
+   "source": [
+    "### Loading a Data File\n",
+    "Loading data from a SLAP2 binary file requires importing the DataFile class from the library. The coorsponding meta data file should be in the same directory with the same file name. If either of these conditions are not met, then an error will be thrown during loading"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "id": "9b219d69-ceaf-4adf-b7be-47e62b1830a2",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from slap2_utils.datafile import DataFile"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "id": "f255492c-0e05-4557-838e-33483bd3070a",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "binaryFilePath = \"3DExampleScan.dat\"\n",
+    "dataFile = DataFile(binaryFilePath)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "ddd7bb0d-dc45-40de-9ba1-4fe2b692fe04",
+   "metadata": {},
+   "source": [
+    "Basic information about the acquisition is stored inside the file header. We can inspect this by accessing the header attribute in the datafile object. This attribute stores information inside a dictionary."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "id": "ec1f034b-00ac-45fc-bcf1-86be4d12cf45",
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "{'firstCycleOffsetBytes': 136.0,\n",
[…]
