+++ setup.cfg
@@ -1,4 +1,0 @@
-# Inside of setup.cfg
-[metadata]
-description-file = README.md
-version = 0.4.1
