+++ MANIFEST.in
@@ -1,3 +1,0 @@
-include requirements.txt
-prune visual_stim
-prune experiments
