### 2025-05-21 383a25d
Updates paper and docs
