+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'SLAP2 Utils: Tools for Processing SLAP2 Data'
+title: 'slap2-utils: Tools for Processing SLAP2 Data'
 tags:
     - Python
     - neuroscience
@@ -31,23 +31,19 @@ bibliography: paper.bib
 
 # Summary
 
-Two-photon microscopy enables the measurements of neural activity in vivo, deep within the brain using fluorescent biosensors. However, capturing neural activity with voltage indicators, fast calcium or neurotransmitter biosensors, or slower indicators across 3D volumes requires sampling speeds that far exceed capabilities of conventional laser-scanning two-photon microscopes. The recently developed Scanned Line Angular Projection-2 (SLAP2) microscope has overcome these limitations by employing novel ultra-fast random-access scanning techniques to record neural activity at kilohertz sampling rates. Unlike traditional raster imaging, where pixels are recorded sequentially by spatial position, SLAP2’s random-access acquisition optimizes the sampling order of regions of interest for speed. These recordings are stored in a complex file structure, which tracks the non-sequentially sampled coordinates. This novel data structure is incompatible with post-imaging analysis tools designed for traditional image formats, such as TIF or CZI files. Here, we present our Python library, `SLAP2 Utils`, to interact with the custom data structure generated from SLAP2. `SLAP2 Utils` allows users to extract neuronal activity directly from the SLAP2 binary files in a Python environment. Additionally, we provide extensions to the SLAP2 microscope control and image acquisition software, providing a framework for user-designed microscope functions during in vivo experiments.  
+Two-photon microscopy enables the measurements of neural activity in vivo, deep within the brain using fluorescent biosensors. However, capturing neural activity with voltage indicators, fast calcium or neurotransmitter biosensors, or slower indicators across 3D volumes requires sampling speeds that far exceed capabilities of conventional laser-scanning two-photon microscopes. The recently developed Scanned Line Angular Projection-2 (SLAP2) microscope has overcome these limitations by employing novel ultra-fast random-access scanning techniques to record neural activity at kilohertz sampling rates. Unlike traditional raster imaging, where pixels are recorded sequentially by spatial position, SLAP2’s random-access acquisition optimizes the sampling order of regions of interest for speed. These recordings are stored in a complex file structure, which tracks the non-sequentially sampled coordinates. This novel data structure is incompatible with post-imaging analysis tools designed for traditional image formats, such as TIF or CZI files. Here, we present our Python library, `slap2-utils`, to interact with the custom data structure generated from SLAP2. `slap2-utils` allows users to extract neuronal activity directly from the SLAP2 binary files in a Python environment.
 
 # Statement of need
 
[…]
