+++ README.md
@@ -1,7 +1,7 @@
-# SLAP2_Utils
+# slap2-utils
 [![Tests](https://github.com/Peter-Hogg/SLAP2_Utils/actions/workflows/python-test.yml/badge.svg)](https://github.com/Peter-Hogg/SLAP2_Utils/actions/workflows/python-test.yml)[![Documentation](https://github.com/Peter-Hogg/SLAP2_Utils/actions/workflows/sphinx.yml/badge.svg)](https://peter-hogg.github.io/SLAP2_Utils/)
 ## Overview
-The `SLAP2_Utils` library is designed to facilitate the reading of SLAP2 (Scanned Line Angular Projection Microscopy version 2) binary files using Python. This utility aims to support researchers and developers working with data from SLAP2 two-photon microscopes by providing an interface to manipulate and analyze these files directly in Python as an alternative to a Matlab-based workflow. The SLAP2 microscope is a commercially available kit from MBF Bioscience (https://www.mbfbioscience.com/products/slap2). The detailed documentation can be found by clicking the documentation icon above `Overview`.
+The `slap2-utils` library is designed to facilitate the reading of SLAP2 (Scanned Line Angular Projection Microscopy version 2) binary files using Python. This utility aims to support researchers and developers working with data from SLAP2 two-photon microscopes by providing an interface to manipulate and analyze these files directly in Python as an alternative to a Matlab-based workflow. The SLAP2 microscope is a commercially available kit from MBF Bioscience (https://www.mbfbioscience.com/products/slap2). The detailed documentation can be found by clicking the documentation icon above `Overview`.
 
 ## Features
 - **Reading SLAP2 Binary Files**: Convert SLAP2 proprietary binary data into accessible formats for Python.
@@ -15,7 +15,7 @@ The `SLAP2_Utils` library is designed to facilitate the reading of SLAP2 (Scanne
 ## Installation
 Install with pip
 ```bash
-pip install SLAP2-UTILS
+pip install slap2-utils
 ```
 
 
