### 2025-05-21 0a08601
Updates paper and docs
