+++ requirements.txt
@@ -11,5 +11,3 @@ notebook
 pyneurotrace
 zarr          
 dask
-napari
-apoc
