+++ paper/paper.bib
@@ -39,27 +39,6 @@
   url = "http://learningsys.org/nips17/assets/papers/paper_16.pdf"
 }
 
-@article{scikit,
-  title = {scikit-image: Image processing in Python},
-  volume = {2},
-  ISSN = {1548-7105},
-  url = {https://doi.org/10.7717/peerj.453},
-  DOI = {10.7717/peerj.453},
-  number = {453},
-  journal = {PeerJ},
-  author = {van der Walt, Stéfan and Schönberger, Johannes L. and Nunez-Iglesias, Juan and Boulogne, François and Warner, Joshua D. and Yager, Neil and Gouillart, Emmanuelle and Yu, Tony and  the scikit-image contributors},
-  year = {2014},
-}
-
-@misc{pytorch,
-  title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library}, 
-  author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Köpf, Andreas and Yang, Edward and DeVito, Zach and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
-  year = {2019},
-  eprint = {1912.01703},
-  archivePrefix = {arXiv},
-  primaryClass = {cs.LG}
-}
-
 @article{numpy,
   title = {Array programming with {NumPy}},
   author = {Harris, Charles R. and Millman,  K. Jarrod and van der Walt, Stéfan and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and Kerkwijk, Marten H. van and Brett,  Matthew and Haldane, Allan and Rio, Jaime Fernandez del and Wiebe, Mark and Peterson, Pearu and Gérard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
@@ -82,26 +61,13 @@
   author = {Collette, Andrew}
 }
 
-@article{scipy,
-  author = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
-            Haberland, Matt and Reddy, Tyler and Cournapeau, David and
-            Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
-            Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
-            Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
-            Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
-            Kern, Robert and Larson, Eric and Carey, C J and
-            Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
-            {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
-            Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
-            Harris, Charles R. and Archibald, Anne M. and
-            Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
-            {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
-  journal = {Nature Methods},
[…]
