+++ paper/paper.md
@@ -47,7 +47,7 @@ To further aid with retrieving traces of neural activity from the random-access 
 
 # SLAP2 Utility Functions and Visualizations 
 
-Several commonly used functions are included for tasks that would be routinely called upon. These utility functions simplify generating the traces from ROIs, cleaning up traces from volumetric recordings, and extracting ROI information in both 2D and 3D. Additional utility functions interact with other file types generated by the microscope, such as extracting stimulus times from stimulus logs stored in H5 files and functions for averaging reference stacks stored as TIF files. Useful plotting functions are also packaged in the library, such as a function that plots ROIs over a reference image. Furthermore, the package comes with a Python-based data viewer that allows for the inspection of activity and ROIs from a given DAT file with a GUI. The data viewer’s features include 3D inspection of ROIs over a reference volume and interactive plots of activity traces. 
+Several commonly used functions are included for tasks that would be routinely called upon. These utility functions simplify generating the traces from ROIs, cleaning up traces from volumetric recordings, and extracting ROI information in both 2D and 3D. Additional utility functions interact with other file types generated by the microscope, such as extracting stimulus times from stimulus logs stored in H5 files and functions for averaging reference stacks stored as TIF files. Useful plotting functions are also packaged in the library, such as a function that plots ROIs over a reference image. Furthermore, a Python-based graphical data viewer is included for interactive inspection of ROIs and activity traces from a given `.DAT` file. The data viewer’s features include 3D inspection of ROIs over a reference volume and interactive plots of activity traces. The viewer uses GPU-accelerated signal processing [@Coleman2024]. This requires the installation of the optional `cupy` requirements to function [@cupy].
 
 # Availability
 
