### 2025-05-28 Peter-Hogg/SLAP2_Utils#3 @Peter-Hogg
> "Functionality: Have the functional claims of the software been confirmed?" In my opinion only some of the claims could be verified. Reading the SLAP file and associated metadata was accomplished through the provided example notebook. There is also a mention of a "Utility" UI. There was no documentation I found for this usage, so I could not verify this claim.

Thank you for this feedback.Thank you for testing the core functionality related to reading SLAP2 files and metadata via the provided example notebook. I appreciate you pointing out the lack of documentation for the Utility UI component.
In response, I have now added documentation for the Utility UI to the repository. This includes usage instructions, screenshots, and an outline of its capabilities.
If there are any specific UI workflows you would like to see illustrated further, I would be happy to expand on those in future updates.

> "Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution." There is a clear list of dependencies. However, some seem superfluous and it's not clear the canonical list of packages. In requirements.txt, notebook is listed as a dependency, but I do not see it's usage anywhere in the library. h5py, PyQT6, matplotlib, vispy, scikit-image, scipy are all listed twice. pyproject.toml lists pyneurotrace but is not listed in requirements.txt. There is both a pyproject.toml and setup.py file. It's not clear which one is used for the PYPI package (which determines the dependencies at install time).

Legacy installation files have been removed, and old personal files have been purged. The installation process has been cleaned up to rely solely on pyproject.toml. The requirements.txt file remains as a reference, and this is now noted clearly at the top of the document.
The notebook package was included so users could run the example Jupyter notebook. If this is considered inappropriate as a required dependency, I can remove it or shift it to an optional group.

>  There also appears to be expected imports that are not stated. E.g.[ zarr, dask](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/imagestacks.py#L2-L9),[ apoc](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/autoROI.py#L2),[ napari](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/autoROI.py#L8), and many others. Furthermore, without pinned requirements in requirements.txt, there are possible pitfalls as some functions appear deprecated. E.g.[ skeletonize_3d](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/slap2_utils/functions/autoROI.py#L12) is no longer a function in[ skimage.morphology](https://scikit-image.org/docs/stable/api/skimage.morphology.html#skimage.morphology.skeletonize).

This has been addressed by modernizing the installation process for the Python library.
Previously missing dependencies such as zarr, dask, apoc, and napari have now been added to the appropriate sections or removed during code clean up. The list of required packages has been audited and updated to reflect actual usage.

> "Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)." There is an example notebook buried in the examples directory. However, I saw no mention of this in the claimed canonical documentation. Note that you can[ include notebooks](https://docs.readthedocs.com/platform/stable/guides/jupyter.html) in Sphinx documentation. It was unclear if the example .dat file was a "real-world" image. I found no documentation related to the UI or its usage in the documentation. If that section is still in progress I would remove it from the paper. It's unclear if[ this](https://github.com/Peter-Hogg/SLAP2_Utils/tree/main/slap2_utils/dataviewer/plotWindows) is the intended module for usage.

I was not previously aware that Jupyter notebooks could be directly integrated into Sphinx documentation. This has now been addressed:
The example notebook has been added to the Sphinx documentation and is now properly linked in the canonical docs.

Documentation for the data viewer UI has also been added, including a description of its purpose and usage.

These updates should make the intended usage of modules and entry points much clearer to users.

> "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?" Documentation exists. However, the docs have errors in their builds. E.g. the imagestacks module does not populate because of a[ build error](https://github.com/Peter-Hogg/SLAP2_Utils/actions/runs/14868419567/job/41750870790). I would encourage the authors to use sphinx-build as it provides a number of[ helpful parameters](https://www.sphinx-doc.org/en/master/man/sphinx-build.html) such as -n, and -W. While quite nitty, many words that appear meant to be in monospace code are italicized due to single ticks vs double. E.g.[ here](https://github.com/Peter-Hogg/SLAP2_Utils/blob/a29f941d76bc8bf9ef28248510bfb17d103b510c/docs/source/modules/acquisitioncontainer.rst?plain=1#L4).

The build error affecting the imagestacks module has been resolved. I've also updated the documentation build process to use sphinx-build with the -n and -W flags, as suggested. This has already helped catch several minor issues and should improve the overall reliability of the docs moving forward.

> "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?" 1 test exists for reading a single datafile and metadata. It ran and passed. I did not find any testing for any other module or function. Coverage of the slap2_utils directory with the test was 13%. In my opinion, this is not acceptable for a package meant to be distributed. E.g. a simple name error in the[ adaptive_thresh](https://github.com/Peter-Hogg/SLAP2_Utils/blob/a29f941d76bc8bf9ef28248510bfb17d103b510c/slap2_utils/filters/filters.py#L77) function uses np but numpy is not imported, meaning doing filters.adaptive_thresh(...) will always result in a 'np' is not defined error.

Thank you for the feedback on test coverage.
Automated testing has been significantly expanded and now covers the majority of the library, with the exception of the data viewer UI. This should greatly improve the robustness and maintainability of the package.

`"A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?" A statement of need exists. The target audience is not directly addressed.`

Thank you for this observation. The target audience is made explicit in the paper by adding the  following sentence to the "Statement of Need" section:
 To overcome this limitation, we developed the Python library slap2-utils, intended for neuroscientists, engineers, and imaging specialists who use SLAP2 microscopes and prefer to analyze their data using Python-based tools.

> "Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?" There are typos and grammatical issues throughout. E.g. line 65 "numpyr" -> "numpy", line 65/66 "memap" -> "[memmap](https://numpy.org/doc/stable/reference/generated/numpy.memmap.html)". OOP and FOV are abbreviated in parentheses but never used again. The definition of "SLAP" appears different between the paper and the repo. In the paper it's "Scanned Line Angular Projection" but in the[ readme](https://github.com/Peter-Hogg/SLAP2_Utils/blob/main/README.md?plain=1#L4) it's "Scanned Line Projection Microscopy". "Utility" is misspelled in the section title, line 83.

Thank you for pointing out these issues. All typos and grammatical errors in the paper have been reviewed and corrected. These changes should improve the clarity and consistency of the manuscript.
General notes

> The name of the tool in the title of the paper is "SLAP2 Utils" and is referred to as such in the paper, however, the package itself is SLAP2-UTILS, and the repo name is "SLAP2_Utils" and that is used throughout the readme file. This is somewhat inconsistent. It would be nice if the package name in the paper matched the Python package which matched the name stated in the readme.

Thank you for highlighting this inconsistency. I’ve reviewed and standardized the naming throughout the paper, README, and documentation.

> The line between personal experiment repo and distributable library appears non-existent. There is an experiments module with what appears to be auto-generated, personalized files?

Thank you for pointing this out. The line between internal experiments and the core distributable library has now been clarified. The paper has also been updated to reflect this as well.

The experiments module previously contained auto-generated and personalized files that were not intended for end users. These have been removed or relocated to a private, lab-specific repository. The public-facing repository now focuses solely on reusable, documented components intended for distribution and community use. This will make maintaining the project easier.
