### 2025-06-18 281b624
fast getLineData
