+++ slap2_utils/datafile.py
@@ -1,14 +1,10 @@
 import os
 import numpy as np
-import mmap
-import scipy.io
-import h5py
 import re
 
 from .subclasses.metadata import MetaData
 from .utils.file_header import load_file_header_v2
-
-from skimage.draw import polygon_perimeter
+from .fast_line_data import fast_get_line_data
 
 
 class DataFile():
@@ -215,7 +211,7 @@ class DataFile():
         #obj.header.referenceTimestamp = first_line_header.timestamp
         return header
 
-    def getLineData(self, lineIndices, cycleIndices, iChannel=None):
+    def getLineData(self, lineIndices, cycleIndices, iChannel=None, useCython=True):
         """Get line data from the data file.
 
         Parameters
@@ -255,29 +251,38 @@ class DataFile():
             raise ValueError("Channel indices must be <= numChannels")
         
         hMemmap = np.memmap(self.datFileName, dtype='int16', mode='r')
-        lineData = []
         
         try:
-            for idx in range(len(lineIndices)):
-                tmpData = np.zeros((self.lineDataNumElements[lineIndices[idx]-1] // int(self.header['numChannels']), len(iChannel)), dtype=np.int16)
-                for ch in range(len(iChannel)):
-                    byteOffsets = [x*2 for x in range(self.lineDataNumElements[lineIndices[idx]-1] // int(self.header['numChannels']))]
-                    byteOffsets = [x + self.lineDataNumElements[lineIndices[idx]-1]//int(self.header['numChannels'])*2*(iChannel[ch]-1) for x in byteOffsets]
-                    byteOffsets = [x + self.lineDataStartIdxs[lineIndices[idx]-1] * 2 for x in byteOffsets]
-                    byteOffsets = [int(x-self.header['firstCycleOffsetBytes']) for x in byteOffsets]
-
-                    cycleIdxs = cycleIndices[idx] - 1
-                    cycleByteOffsets = self.header['firstCycleOffsetBytes'] + cycleIdxs*self.header['bytesPerCycle']
-                    cycleSampleOffsets = int(cycleByteOffsets//2)
-
-                    sampleOffsets = np.array([int(x//2) + cycleSampleOffsets for x in byteOffsets], dtype=np.uint64)
-
-                    tmpData[:, ch] = hMemmap[sampleOffsets-1]
-                
-                lineData.append(tmpData)
+            if useCython:
+                lineData = fast_get_line_data(
+                    hMemmap,
+                    self.lineDataNumElements,
+                    self.lineDataStartIdxs,
+                    int(self.header['numChannels']),
+                    int(self.header['firstCycleOffsetBytes']),
+                    int(self.header['bytesPerCycle']),
+                    lineIndices,
+                    cycleIndices,
+                    iChannel
+                )
+            else:
+                lineData = []
+                for idx in range(len(lineIndices)):
+                    tmpData = np.zeros((self.lineDataNumElements[lineIndices[idx]-1] // int(self.header['numChannels']), len(iChannel)), dtype=np.int16)
+                    for ch in range(len(iChannel)):
+                        byteOffsets = [x*2 for x in range(self.lineDataNumElements[lineIndices[idx]-1] // int(self.header['numChannels']))]
+                        byteOffsets = [x + self.lineDataNumElements[lineIndices[idx]-1]//int(self.header['numChannels'])*2*(iChannel[ch]-1) for x in byteOffsets]
+                        byteOffsets = [x + self.lineDataStartIdxs[lineIndices[idx]-1] * 2 for x in byteOffsets]
+                        byteOffsets = [int(x-self.header['firstCycleOffsetBytes']) for x in byteOffsets]
[…]
