### 2025-06-19 df1713f
add numpy version
