### 2025-06-23 119e8af
acquisitionContainer: add compatibility for new pixelDilation format
