### 2025-06-25 b0c67a8
keep only numpy version
