### 2025-06-25 5ee5117
Revert "keep only numpy version"
This reverts commit b0c67a8bf274dbf0799999aaa646b00dace593bb.
