+++ requirements.txt
@@ -11,3 +11,4 @@ notebook
 pyneurotrace
 zarr          
 dask
+cython
