### 2025-06-25 672ce6e
build with hatch. try compiled fast_get_line_data first, use numpy if it fails
