+++ slap2_utils/datafile.py
@@ -4,7 +4,6 @@ import re
 
 from .subclasses.metadata import MetaData
 from .utils.file_header import load_file_header_v2
-from .fast_line_data import fast_get_line_data
 
 
 class DataFile():
@@ -255,7 +254,8 @@ class DataFile():
         hMemmap = np.memmap(self.datFileName, dtype='int16', mode='r')
         
         try:
-            if method == "Cython":
+            try:
+                from .fast_line_data import fast_get_line_data
                 lineData = fast_get_line_data(
                     hMemmap,
                     self.lineDataNumElements,
@@ -267,7 +267,7 @@ class DataFile():
                     cycleIndices,
                     iChannel
                 )
-            elif method == "Numpy":
+            except ImportError:
                 lineData = []
                 for idx in range(len(lineIndices)):
                     samples_per_channel = int(self.lineDataNumElements[lineIndices[idx]-1]) // int(self.header['numChannels'])
@@ -277,21 +277,6 @@ class DataFile():
                     for ch in range(len(iChannel)):
                         offset = samples_per_channel * (iChannel[ch]-1) + precomputed_offset - 1
                         tmpData[:, ch] = hMemmap[offset:offset + samples_per_channel]
-                    lineData.append(tmpData)
-            else:
-                lineData = []
-                for idx in range(len(lineIndices)):
-                    tmpData = np.zeros((self.lineDataNumElements[lineIndices[idx]-1] // int(self.header['numChannels']), len(iChannel)), dtype=np.int16)
-                    for ch in range(len(iChannel)):
-                        byteOffsets = [x*2 for x in range(self.lineDataNumElements[lineIndices[idx]-1] // int(self.header['numChannels']))]
-                        byteOffsets = [x + self.lineDataNumElements[lineIndices[idx]-1]//int(self.header['numChannels'])*2*(iChannel[ch]-1) for x in byteOffsets]
-                        byteOffsets = [x + self.lineDataStartIdxs[lineIndices[idx]-1] * 2 for x in byteOffsets]
-                        byteOffsets = [int(x-self.header['firstCycleOffsetBytes']) for x in byteOffsets]
-                        cycleIdxs = cycleIndices[idx] - 1
-                        cycleByteOffsets = self.header['firstCycleOffsetBytes'] + cycleIdxs*self.header['bytesPerCycle']
-                        cycleSampleOffsets = int(cycleByteOffsets//2)
-                        sampleOffsets = np.array([int(x//2) + cycleSampleOffsets for x in byteOffsets], dtype=np.uint64)
-                        tmpData[:, ch] = hMemmap[sampleOffsets-1]
                     lineData.append(tmpData)
         finally:
             if hasattr(hMemmap, '_mmap'):
