+++ slap2_utils/functions/tracefunctions.py
@@ -75,7 +75,7 @@ def cleanVolumeTrace(datafile, zId, rawTrace):
     """
 
     # Extract the slice index information from the datafile metadata
-    sliceIdx = datafile.metaData.AcquisitionContainer.ParsePlan['acqParsePlan']['sliceIdx']
+    sliceIdx = datafile.metaData.AcquisitionContainer.AcquisitionPlan['acqParsePlan']['sliceIdx']
     dataIndex = []
     
     # Adjust zId for Python's 0-based indexing
