+++ slap2_utils/datafile.py
@@ -17,7 +17,7 @@ class DataFile():
 
             Attributes
             ----------
-            MAGIC_NUMBER : ``np.uint32``
+            MAGIC_NUMBER : np.uint32
                 A constant used to validate the data file.
             filename : str
                 The name of the data file.
@@ -27,7 +27,7 @@ class DataFile():
                 The name of the data file with the '.dat' extension.
             rawData : list
                 A list to store the raw data from the file.
-            metaData : ``MetaData``
+            metaData : MetaData
                 An instance of the MetaData class to handle metadata.
             StreamId : None
                 A placeholder for stream ID.
@@ -125,54 +125,32 @@ class DataFile():
         # Subfunction of loading parse plan
         def load_parse_plan(self, metaData):   
             def filter_z_pixel_replacement_maps(z_maps):
-                # Filter out redundant mappings where source and destination indices are the same
-                filtered_maps = []
-                for slice_map in z_maps:
-                    redundant_mask = slice_map[0] == slice_map[1]
-                    filtered_maps.append(slice_map[:,~redundant_mask])
-                return filtered_maps
+                # Using list comprehension for simplified logic
+                return [list(filter(lambda x: x[0] != x[1], map_)) for map_ in z_maps]
 
-            if not metaData.AcquisitionContainer.newFormat:
-                fastz = metaData.AcquisitionContainer.ParsePlan['zs'][:]
-                final_fastz = [-10203] #initiating with a value thats never found
-                for i in fastz:
-                    final_fastz.append(i[0])
+            fastz = metaData.AcquisitionContainer.AcquisitionPlan['zs'][:]
+            final_fastz = [-10203] #initiating with a value thats never found
+            for i in fastz:
+                final_fastz.append(i[0])
 
-                self.fastZs = final_fastz
+            self.fastZs = final_fastz
 
-                #Check if it breaks
-                self.lineSuperPixelZIdxs = metaData.AcquisitionContainer.ParsePlan['acqParsePlan']['sliceIdx']
-                self.lineSuperPixelIDs = [x.T for x in metaData.AcquisitionContainer.ParsePlan['acqParsePlan']['superPixelID']]
-                
-                self.zPixelReplacementMaps = metaData.AcquisitionContainer.ParsePlan['pixelReplacementMaps']
-                
-                #Using list comprehension for simplified logic
-                self.lineNumSuperPixels = [len(ids) for ids in self.lineSuperPixelIDs]
-                self.lineFastZIdxs = np.zeros(len(self.lineSuperPixelZIdxs))
-                for lineIdx in range(len(self.lineSuperPixelZIdxs)):
-                    lineZIdxs_=self.lineSuperPixelZIdxs[lineIdx]
+            #Check if it breaks
+            self.lineSuperPixelZIdxs = metaData.AcquisitionContainer.AcquisitionPlan['acqParsePlan']['sliceIdx']
+            self.lineSuperPixelIDs = metaData.AcquisitionContainer.AcquisitionPlan['acqParsePlan']['superPixelID']
             
-                    if len(lineZIdxs_) != 1:
-                        self.lineFastZIdxs[lineIdx] = 0
-                    else:
-                        self.lineFastZIdxs[lineIdx] = lineZIdxs_[0][0] + 1
-            else:
-                print('Using new metadata format...')
-
[…]
