### 2025-07-13 aa6ffca
Update paper.bib
