+++ paper/paper.bib
@@ -22,15 +22,6 @@
   url = {https://www.mbfbioscience.com/products/slap2}
 }
 
-@article{phasecross,
-  title = {The phase correlation image alignment method},
-  journal = {Proceeding of IEEE International Conference on Cybernetics and Society},
-  author = {Kuglin, Charles and Hines, D. C.},
-  year = {1975},
-  pages = {163-165}
-}
-
-
 @inproceedings{cupy,
   author = "Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman",
   title = "CuPy: A NumPy-Compatible Library for NVIDIA GPU Calculations",
