### 2025-07-17 267a438
Update Jerry.md
