### 2025-07-17 #8054 @Peter-Hogg
I have checked other names. Made a new release (v0.6.1) and made a Zenodo release (https://doi.org/10.5281/zenodo.16045225).

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.16045225.svg)](https://doi.org/10.5281/zenodo.16045225)
