### 2025-07-24 #8054 @Kevin-Mattheus-Moerman
@Peter-Hogg can you set the archive listed version to `0.6.1`
