### 2025-08-20 #8054 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
